refactor(telegram): split lane delivery modules
(cherry picked from commit 269cc22b61d5de1f8a93a3bc22165a82ac21957c)
